+++ src/site/markdown/changelog.md
@@ -4,6 +4,8 @@
 
 - public (static) scopes (see _ticino-static_ class net.micwin.ticino.events.PublicScopes)
 
+- ticino.events.EventScope gets method_ dispatchNoMatterWhat_ that reaches all receiver no matter if a receiver before threw an exception.
+
 ## 0.3.6 
 
 - removed some package issues
@@ -14,3 +16,8 @@
 - Cleanup
 - introduced ticino context
 - gone java 1.8
+
+# open issues
+
+- EventScope.dispatch -> dispatchVote
+- EventScope.dispatchNoMatterWhat -> dispatchAll